removed unused stories versions, and made them displaying real image from the S3 bucket
