+++ packages/api/server.js
@@ -109,9 +109,9 @@ app.get('/api/getImageForStory', visitorHandlers.getImageForStory.bind({amazon: 
 app.get('/api/event', visitorHandlers.volunteerFormGetEventInfo.bind({dbClient: client}))
 app.post('/api/form', visitorHandlers.volunteerFormSubmit.bind({dbClient: client}))
 app.get('/api/eventCalendar', visitorHandlers.eventCalendar.bind({dbClient: client}))
-app.get('/api/displayStory', visitorHandlers.displayStory.bind({dbClient: client}))
+app.get('/api/displayStory', visitorHandlers.displayStory.bind({dbClient: client, amazon: AWS}))
 app.get('/api/countStory', visitorHandlers.countStory.bind({dbClient: client}))
-app.get('/api/getOneStory', visitorHandlers.getOneStory.bind({dbClient: client}))
+app.get('/api/getOneStory', visitorHandlers.getOneStory.bind({dbClient: client, amazon: AWS}))
 app.get('/api/content', visitorHandlers.getContent.bind({dbClient: client}))
 app.get('/api/getCalendarFAQ', visitorHandlers.getCalendarFAQ.bind({dbClient: client}))
