+++ packages/api/visitor.js
@@ -62,7 +62,7 @@ homeImages = function(req, res) {
           Bucket: bucket,
           Key: key,
         }))
-      }   
+      }
     })
     res.send(imageUrls)
   })
@@ -100,7 +100,7 @@ volunteerFormGetEventInfo = function(req, res) {
         min_vegan_servings: Math.ceil(Math.floor(category.min_servings/3) / 10) * 10,
         food: category.food,
         min_vegan: category.min_vegan,
-        real_vegan: 0, 
+        real_vegan: 0,
         servings: 0
       })
       category.submissions.forEach(sub => {
@@ -172,7 +172,7 @@ volunteerFormSubmit = function(req, res) {
 
 eventCalendar = function(req, res) {
   let client = this.dbClient
-  collection = client.db('events-form').collection('events') 
+  collection = client.db('events-form').collection('events')
   collection.find({}).toArray((err, docs) => {
     if(err) {
       console.log(err, 'Error trying to find document')
@@ -216,7 +216,6 @@ displayStory = function(req, res) {
   let skips = (req.query.page - 1) * 12
   let client = this.dbClient
   collection = client.db('stories_example1').collection('published')
-  console.log('publish page ' + req.query.page)
   collection.find().sort({ _id: -1 }).skip(skips).limit(12).toArray((err, docs) => {
     if (err) {
       console.log(err, 'Error trying to find published Stories')
@@ -231,6 +230,14 @@ displayStory = function(req, res) {
       })
       return
     }
+    let AWS = this.amazon
+    const s3 = new AWS.S3({
+      endpoint: new AWS.Endpoint(process.env.S3_BUCKET),
+      s3ForcePathStyle: true,
+      accessKeyId: process.env.S3_ACCESS_KEY,
+      secretAccessKey: process.env.S3_SECRET_ACCESS_KEY
+    })
+    const bucket = 'beautiful-portland-carousel-photos'
     let response_data = []
     docs.map((pubStory) => {
       var publishObj = new Object()
@@ -239,9 +246,25 @@ displayStory = function(req, res) {
       publishObj.title = pubStory.title
       publishObj.hook = pubStory.hook
       publishObj.content = pubStory.content
-      publishObj.public_status = pubStory.public_status
-      publishObj.postPhotoName = pubStory.postPhotoName
-      response_data.push(pubStory)
+      if(pubStory.postPhotoName=== 'No Photo'){
+        publishObj.imageUrl = 'notFound'
+        response_data.push(publishObj)
+      }else{
+        let key = 'storyPhotos/' + pubStory.postPhotoName
+        try {
+          let url = s3.getSignedUrl('getObject', {Bucket: bucket, Key: key})
+          if(url.indexOf('undefined') === -1) {
+            publishObj.imageUrl = url
+          }
+          else {
+            console.log('File not found!')
+            publishObj.imageUrl = 'notFound'
+          }
+        } catch (err) {
+          console.log('ERROR could not locate file : ' + JSON.stringify(err))
+        }
+        response_data.push(publishObj)
+      }
     })
     res.send({
       status: 'SUCCESS',
@@ -257,7 +280,6 @@ countStory = async function(req, res) {
   countObj.publishCount = await collection.countDocuments({})
   let response_data = []
   response_data.push(countObj)
-  console.log(response_data)
   res.send({
     status: 'SUCESS',
     count_info: JSON.stringify(response_data)
@@ -281,6 +303,14 @@ getOneStory = function(req, res) {
       })
       return
     }
+    let AWS = this.amazon
+    const s3 = new AWS.S3({
+      endpoint: new AWS.Endpoint(process.env.S3_BUCKET),
+      s3ForcePathStyle: true,
+      accessKeyId: process.env.S3_ACCESS_KEY,
+      secretAccessKey: process.env.S3_SECRET_ACCESS_KEY
+    })
+    const bucket = 'beautiful-portland-carousel-photos'
     let response_data = []
     docs.map((pubStory) => {
       var publishObj = new Object()
@@ -289,8 +319,24 @@ getOneStory = function(req, res) {
       publishObj.title = pubStory.title
       publishObj.hook = pubStory.hook
       publishObj.content = pubStory.content
-      publishObj.public_status = pubStory.public_status
-      response_data.push(pubStory)
+      if(pubStory.postPhotoName=== 'No Photo'){
+        publishObj.imageUrl = 'notFound'
+        response_data.push(publishObj)
+      }else{
+        let key = 'storyPhotos/' + pubStory.postPhotoName
+        try {
+          let url = s3.getSignedUrl('getObject', {Bucket: bucket, Key: key})
+          if(url.indexOf('undefined') === -1) {
+            publishObj.imageUrl = url
+          }
+          else {
+            publishObj.imageUrl = 'notFound'
+          }
+        } catch (err) {
+          console.log('ERROR could not locate file : ' + JSON.stringify(err))
+        }
+        response_data.push(publishObj)
+      }
     })
     res.send({
       status: 'SUCCESS',
@@ -345,7 +391,7 @@ getContent= function(req, res) {
     })
   }
 }
-  
+
 getCalendarFAQ = function(req, res) {
   let client = this.dbClient
   collection = client.db('events-form').collection('frequently-asked-questions')
