+++ packages/ui/src/Components/Stories/Story.js
@@ -1,82 +1,96 @@
 import React, { Component } from 'react'
-import { Header, Modal, Image, Container, Icon } from 'semantic-ui-react'
-import Moment from 'moment'
-import Axios from 'axios'
+import { Card, Image, Icon, Button, Modal } from 'semantic-ui-react'
+import { Route, Link } from 'react-router-dom'
+import ReadMore from './ReadMore'
 import '../Stylesheets/PublishStories.css'
-import { Dropdown } from 'semantic-ui-react'
+import LogoPlaceHolder from './../../logoPhotos/bpdx_story_place_holder.png'
+import Moment from 'moment'
 
 // Sitewide Text Styles
 const paragraphStyles = {
   fontFamily: 'Nunito',
-  color: '#000b91', 
+  color: '#000b91',
   fontSize: 'calc(14px + (26 - 14) * ((100vw - 300px) / (1300)))',
   lineHeight: 'calc(1.3em + (1.5 - 1.2) * ((100vw - 300px)/(1300)))'
-} 
-
-const headerStyles = {
-  fontFamily: 'Nunito',
-  color: '#000b91', 
-  fontSize: 'calc(20px + (42 - 20) * ((100vw - 300px) / (1300)))',
-  lineHeight: 'calc(1.3em + (1.5 - 1.2) * ((100vw - 300px)/(1300)))}'
 }
 
+// const headerStyles = {
+//   fontFamily: 'Nunito',
+//   color: '#000b91',
+//   fontSize: 'calc(20px + (42 - 20) * ((100vw - 300px) / (1300)))',
+//   lineHeight: 'calc(1.3em + (1.5 - 1.2) * ((100vw - 300px)/(1300)))}'
+// }
+
 const dropdownStyles = {
-  fontFamily: 'Nunito', 
-  color: '#000b91', 
+  fontFamily: 'Nunito',
+  color: '#000b91',
   fontSize: 'calc(12px + (20 - 12) * ((100vw - 300px) / (1300)))',
   lineHeight: 'calc(1.3em + (1.5 - 1.2) * ((100vw - 300px)/(1300)))'
 }
 
-export default class PublishStories extends Component {
+export default class Story extends Component {
   constructor(props){
     super(props)
     this.state = {
-      publishStory: [],
-      date: '',
-      title: this.props.title
+      date: Moment(this.props.publish.edited_timestamp).format('YYYY/MM/D, HH:mm A'),
+      year: Moment(this.props.publish.edited_timestamp).format('YYYY'),
+      month: Moment(this.props.publish.edited_timestamp).format('MMMM'),
+      version: this.props.version,
+      url: '/Stories/'+this.props.url,
+      open: false
     }
   }
 
-  componentDidMount() {
-    let path = '/api/getOneStory?title=' + this.props.title
-    console.log(this.props.title)
-    Axios.get(path)
-      .then((response) => {
-        let tempPubStory = JSON.parse(response.data.published_info)
-        console.log(response)
-        this.setState({
-          publishStory: tempPubStory[0],
-          date: Moment(tempPubStory[0].edited_timestamp).format('YYYY/MM/D, HH:mm A')
-        })
-        console.log(this.state.publishStory)
-      })
-      .catch((err) => {
-        console.log(err, 'Try again.')
-      })
+  show = dimmer => () => {
+    this.setState({ dimmer, open: true })
   }
 
-    close = () => this.setState({ open: false })
+  close = () => {
+    this.setState({ open: false })
+  }
 
-    render () {
-      return (
-        <div>
-          <Modal.Content>
-            <Modal.Description>
-              <Image size='massive' src='https://www.visittheusa.com/sites/default/files/styles/hero_m_1300x700/public/images/hero_media_image/2017-04/048df232dcf31e992cb9143f695f416e.jpeg?itok=8AQcj59p'/><br/>
-              <Container text>
-                <Header as='h1' style= {paragraphStyles}><div className='capitalize'>{this.state.title}</div>
-                  <Header.Subheader style={dropdownStyles}><Icon name='calendar'/>Last Edited: {this.state.date}</Header.Subheader>
-                </Header>
-                <Header as='h3' style={paragraphStyles}><i>{this.state.publishStory.hook}</i></Header>
-                <p style={paragraphStyles}>{this.state.publishStory.content}</p>
-                <Image src='https://react.semantic-ui.com/images/wireframe/paragraph.png' /><br/>
-                <Image src='https://react.semantic-ui.com/images/wireframe/paragraph.png' /><br/>
-                <Image src='https://react.semantic-ui.com/images/wireframe/paragraph.png' /><br/>
-                <Image src='https://react.semantic-ui.com/images/wireframe/paragraph.png' />
-              </Container>
-            </Modal.Description>
-          </Modal.Content>
-        </div>
-      )
-    }
+  render () {
+    const { open, dimmer } = this.state
+    const imageUrl = this.props.publish.imageUrl
+
+    return (
+      <Card raised>
+        <div className='outer-div'>{imageUrl==='notFound' ? (
+          <Image className='card-image' fluid src={LogoPlaceHolder} />
+        ):(
+          <Image className='card-image' fluid src={this.props.publish.imageUrl} />
+        )}</div>
+        <Card.Content>
+          <Card.Header to='/as'>
+            <div className='capitalize'>{this.props.publish.title}</div>
+          </Card.Header>
+
+          <Card.Meta>
+            <Icon name='calendar'/>{' '}{this.state.date} edited
+          </Card.Meta>
+
+          <Card.Description>
+            <p>{this.props.publish.hook}</p>
+          </Card.Description>
+          <br/><br/>
+
+          <Link to={this.state.url} title={this.props.publish.title}>
+            <Button size='small' floated='left' onClick={this.show('blurring')} style={{position: 'absolute', bottom: '10px', boxShadow: '3px 3px #888888', paragraphStyles, backgroundColor:'#a8efae'}}>READ MORE</Button>
+          </Link>
+
+          <Modal dimmer={dimmer} open={open} onClose={this.close}>
+            <Route
+              path={this.state.url}
+              render={(props)=><ReadMore title={this.props.publish.title}/>}
+            />
+            <Modal.Actions>
+              <Link to='Stories'>
+                <Button size='large' onClick={this.close} style={{backgroundColor: '#a8efae'}}>Close</Button>
+              </Link>
+            </Modal.Actions>
+          </Modal>
+        </Card.Content>
+      </Card>
+    )
+  }
 }
